SELinux: remove crazy contortions around proc
We check if the fsname is proc and if so set the proc superblock security
struct flag.  We then check if the flag is set and use the string 'proc'
for the fsname instead of just using the fsname.  What's the point?  It's
always proc...  Get rid of the useless conditional.

Change-Id: Ic4cd1d3507ae5ab3346746d391070e3c899da59c
Signed-off-by: Eric Paris <eparis@redhat.com>
Signed-off-by: Marc Babad <blitzkrieg.marc02@gmail.com>